Added analytics events for login and verify
